+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added image dimensions to mini gallery
 * Added PGP and S/MIME key indicators to address bubbles
 * Added delete existing accounts option to import dialog
+* Added clear all data settings tile
 * Small improvements and minor bug fixes
 * Updated translations